refactor: extract route utility methods
